requirements.txt
Django==4.1
djangorestframework==3.13.1
# djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
PyJWT==2.4.0
